CXAN-71 Rename models-core to models-annotations, move TimeAdapter to api-core
