+++ models-parcelable/build.gradle
@@ -50,7 +50,7 @@ android {
 }
 
 dependencies {
-    implementation project(':models-core')
+    implementation project(':models-annotations')
 
     implementation libs.moshi
     ksp libs.moshi.codegen
