+++ settings.gradle
@@ -5,7 +5,7 @@ include ':vimeo-networking',
         ':auth',
         ':api-core',
         ':request',
-        ':models-core',
+        ':models-annotations',
         ':models',
         ':models-parcelable',
         ':models-serializable',
